Remove view API from app.py.
The API for the views has been put in a Python module called 'view_api'.
